Own files for RegistryFixtures
